refactor(ops): give the Railway scripts one CLI surface and one record model (#6142)
Closes the structural findings the review left, rather than recording them.

Three scripts talk to the same production project and had begun carrying
private copies of the same things. The duplication was not stylistic: the
trigger's own copy of the deployment recency sort had ALREADY diverged
from the drift check's — `Date.parse(x?.createdAt ?? 0)` yields NaN on a
malformed timestamp and leaves the sort order undefined, where the check's
collapsed NaN to "oldest". Both decide "which deployment is this service
running" from that sort, so the two surfaces could disagree about one
service: the check reporting one the trigger considers handled, or the
trigger deploying one the check calls current.

Two new modules, both imported rather than copied:

- scripts/railway-deployments.mjs — what a deployment RECORD means, no
  I/O: the status sets, createdAtMs, orderByRecency, newestRunning.
- scripts/railway-cli.mjs — the CLI calls themselves: runRailway,
  readDeployments, readEnvironmentConfig (fail-closed), readRepositoryServices,
  mapWithConcurrency, readArgument, isRepositoryService.

normalizeRootDirectory and ROOT_DIRECTORY_BY_DEPLOY_MODE now live once, in
the pure closure module, with the audit importing them — which also
removes the map-equality test that existed only because there were two.
The audit, the drift check and the trigger re-export what their existing
importers use, so nothing outside had to change.

Also from the review:

- The trigger gets its OWN deployment window (200, not the drift check's
  50). "Never re-trigger a build Railway already ran and FAILED" holds
  only while that record is still in the window, and a 15-minute cron
  pushes it out of 50 records in half a day — after which the trigger
  would retry a genuinely failing build and bury the alarm in a loop.
- Exit code split by who can act. One unreadable service is ordinary
  third-party rot: it is warned about, the other services are still
  planned, and the drift check alarms independently if that one really is
  behind — so it no longer reds a scheduled run, because a routinely-red
  workflow stops being read. EVERY service unreadable still fails: that is
  an auth or connectivity failure wearing per-service clothing, and
  planning nothing while reporting success is the silent no-op this script
  exists to remove.
- The fleet-wide measurement is stated once, in the solutions document
  that owns it, and referenced from the scripts instead of restated in
  five places where the numbers would rot independently.
- The run line now prints reads= and elapsed=, which is what the pending
  API-fan-out question needs and what nothing was recording.

Verified live after the move: the trigger dry-runs, the audit passes, and
an unmerged SHA is still refused.
